* changed last_calibration_high_transition_            last_calibration_low_transition_  to int32_t from double.  * added unadvertise calls to mechanism control.
git-svn-id: https://svn.code.sf.net/p/personalrobots/code/pkg/trunk@4720 f5854215-dd47-0410-b2c4-cdd35faa7885

+++ mechanism/hardware_interface/include/hardware_interface/hardware_interface.h
@@ -63,8 +63,8 @@ public:
   double velocity_;
 
   bool calibration_reading_;
-  double last_calibration_high_transition_;
-  double last_calibration_low_transition_;
+  int32_t last_calibration_high_transition_;
+  int32_t last_calibration_low_transition_;
 
   bool is_enabled_;
   bool run_stop_hit_;
